first day of beta testing fixes

+++ app/src/main/java/com/example/cissy/esm/Sample.java
@@ -3,6 +3,9 @@ package com.example.cissy.esm;
 import android.app.Service;
 import android.content.Context;
 import android.content.pm.PackageManager;
+import android.graphics.Typeface;
+import android.text.Spannable;
+import android.text.SpannableStringBuilder;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -10,6 +13,7 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.FrameLayout;
 import android.widget.LinearLayout;
+import android.widget.RadioButton;
 import android.widget.RadioGroup;
 import android.widget.TextView;
 
@@ -28,6 +32,7 @@ public class Sample {
     private int type;
     private int sampleTime = 0;
     private String packageName;
+    private String promptType;
 
     public Sample(int type) {
         this.type = type;
@@ -51,9 +56,13 @@ public class Sample {
         }
     }
 
+    public void setPromptType(String type) { this.promptType = type; }
+
     public void setPackageName(String name) { this.packageName = name; }
 
     public String getPackageName() { return this.packageName; }
+
+    public String getPromptType() { return this.promptType; }
 
     public void setSampleTime(int time) { this.sampleTime = time; }
 
@@ -80,21 +89,30 @@ public class Sample {
     public View getPopup(final DetectAppsService service) {
         final Sample thisSample = this;
         LayoutInflater li = (LayoutInflater) service.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
-        final LinearLayout popup;
+        final LinearLayout popup = (LinearLayout) li.inflate(R.layout.affect_before, null);
+        TextView prompt = (TextView) popup.findViewById(R.id.textView);
+        final SpannableStringBuilder str;
 
         switch (this.getType()) {
             case BEFORE:
-                popup = (LinearLayout) li.inflate(R.layout.affect_before, null);
+                str = new SpannableStringBuilder("How were you feeling right before launching this app?");
+                str.setSpan(new android.text.style.StyleSpan(Typeface.ITALIC), 20, 33, Spannable.SPAN_EXCLUSIVE_EXCLUSIVE);
+                prompt.setText(str);
                 break;
             case DURING:
-                popup = (LinearLayout) li.inflate(R.layout.affect_during, null);
+                str = new SpannableStringBuilder("How are you feeling right now?");
+                str.setSpan(new android.text.style.StyleSpan(Typeface.ITALIC), 20, 28, Spannable.SPAN_EXCLUSIVE_EXCLUSIVE);
+                prompt.setText(str);
                 break;
             case AFTER:
-                popup = (LinearLayout) li.inflate(R.layout.affect_during, null);
+                str = new SpannableStringBuilder("How are you feeling right now?");
+                str.setSpan(new android.text.style.StyleSpan(Typeface.ITALIC), 20, 28, Spannable.SPAN_EXCLUSIVE_EXCLUSIVE);
+                prompt.setText(str);
                 break;
             default:
-                popup = null;
+                prompt.setText(null);
         }
+        this.setPromptType("affect_scale");
         Button submit = (Button) popup.findViewById(R.id.btnSubmit);
 
         submit.setOnClickListener(new View.OnClickListener() {
@@ -111,7 +129,12 @@ public class Sample {
                 int idx2 = eGroup.indexOfChild(radioButton2);
 
                 if (idx >= 0 && idx2 >= 0) {
-                    DetectAppsService.sendEvent(service, thisSample.getCategory(), "affect_scale", ";happiness="+idx+";energy="+idx2+";app="+thisSample.getPackageName()+";duration="+thisSample.getSampleTime());
+                    Event event = new Event.EventBuilder(1, System.currentTimeMillis(), thisSample.getPackageName())
+                            .valence(Integer.toString(idx))
+                            .arousal(Integer.toString(idx2))
+                            .durationBefore(thisSample.getSampleTime())
+                            .build();
+                    //DetectAppsService.sendEvent(service, thisSample.getCategory(), "affect_scale", event);
                     service.removeView(popup);
 
                     Random rand = new Random();
@@ -121,16 +144,26 @@ public class Sample {
                         final EditText answer = (EditText) followUp.findViewById(R.id.editText);
 
                         service.addView(followUp);
+                        thisSample.setPromptType("affect_text");
                         Button submit2 = (Button) followUp.findViewById(R.id.btnSubmit);
 
                         submit2.setOnClickListener(new View.OnClickListener() {
                             @Override
                             public void onClick(View v) {
-                                DetectAppsService.sendEvent(service, thisSample.getCategory(), "affect_text", ";response="+answer.getText()+";app="+thisSample.getPackageName()+";duration="+thisSample.getSampleTime());
-                                service.removeView(followUp);
-                                if (thisSample.getType() == Sample.AFTER) {
-                                    final View followUp = thisSample.getMeaningfulnessLikert(service);
-                                    service.addView(followUp);
+                                String response = answer.getText().toString();
+                                if (response.length() >= 75) {
+                                    //DetectAppsService.sendEvent(service, thisSample.getCategory(), thisSample.getPromptType(), ";response="+answer.getText()+";app="+thisSample.getPackageName()+";duration="+thisSample.getSampleTime());
+                                    service.removeView(followUp);
+                                    if (thisSample.getType() == Sample.AFTER) {
+                                        final View followUp = thisSample.getMeaningfulnessLikert(service);
+                                        service.addView(followUp);
+                                        thisSample.setPromptType("meaningfulness_scale");
+                                    }
+                                    else {
+                                        final View followUp = thisSample.getPurpose(service);
+                                        service.addView(followUp);
+                                        thisSample.setPromptType("purpose_mc");
+                                    }
                                 }
                             }
                         });
@@ -138,6 +171,12 @@ public class Sample {
                     else if (thisSample.getType() == Sample.AFTER) {
                         final View followUp = thisSample.getMeaningfulnessLikert(service);
                         service.addView(followUp);
+                        thisSample.setPromptType("meaningfulness_scale");
+                    }
+                    else {
+                        final View followUp = thisSample.getPurpose(service);
+                        service.addView(followUp);
+                        thisSample.setPromptType("purpose_mc");
                     }
                 }
             }
@@ -168,13 +207,60 @@ public class Sample {
         return popup;
     }
 
+    public View getPurpose(final DetectAppsService service) {
+        final Sample thisSample = this;
+        LayoutInflater li=(LayoutInflater) service.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
+        final LinearLayout popup = (LinearLayout) li.inflate(R.layout.purpose_mc, null);
+        Button submit = (Button) popup.findViewById(R.id.btnSubmit);
+
+        submit.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                RadioGroup mGroup = (RadioGroup) popup.findViewById(R.id.purpose_group);
+                int checkedM = mGroup.getCheckedRadioButtonId();
+                RadioButton radioButton = (RadioButton) mGroup.findViewById(checkedM);
+                String purpose = radioButton.getText().toString();
+
+                //DetectAppsService.sendEvent(service, thisSample.getCategory(), thisSample.getPromptType(), ";purpose="+purpose+";app="+thisSample.getPackageName()+";duration="+thisSample.getSampleTime());
+                service.removeView(popup);
+
+                if (purpose.equals("Communicating / interacting")) {
+                    final View closenessFollowUp = thisSample.getCloseness(service);
+                    service.addView(closenessFollowUp);
+                    thisSample.setPromptType("closeness_scale");
+                }
+            }
+        });
+        return popup;
+    }
+
+    public View getCloseness(final DetectAppsService service) {
+        final Sample thisSample = this;
+        LayoutInflater li=(LayoutInflater) service.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
+        final LinearLayout popup = (LinearLayout) li.inflate(R.layout.closeness_scale, null);
+        Button submit = (Button) popup.findViewById(R.id.btnSubmit);
+
+        submit.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                RadioGroup mGroup = (RadioGroup) popup.findViewById(R.id.closeness_group);
+                int checkedM = mGroup.getCheckedRadioButtonId();
+                RadioButton radioButton = (RadioButton) mGroup.findViewById(checkedM);
+                int idx = mGroup.indexOfChild(radioButton);
+                //DetectAppsService.sendEvent(service, thisSample.getCategory(), thisSample.getPromptType(), ";closeness="+idx+";app="+thisSample.getPackageName()+";duration="+thisSample.getSampleTime());
+                service.removeView(popup);
+            }
+        });
+        return popup;
+    }
+
     public View getMeaningfulnessLikert(final DetectAppsService service) {
         final Sample thisSample = this;
         LayoutInflater li=(LayoutInflater) service.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
         final LinearLayout popup = (LinearLayout) li.inflate(R.layout.meaningfulness_scale, null);
         TextView prompt = (TextView) popup.findViewById(R.id.textView);
 
-        prompt.setText("How meaningful was using "+service.getSampledApp()+" just now?");
+        prompt.setText("You just used "+service.getSampledApp()+".\n\nHow much do you feel like you have spent your time on something meaningful?");
 
         Button submit = (Button) popup.findViewById(R.id.btnSubmit);
 
@@ -185,7 +271,7 @@ public class Sample {
                 int checkedM = mGroup.getCheckedRadioButtonId();
                 View radioButton = mGroup.findViewById(checkedM);
                 int idx = mGroup.indexOfChild(radioButton);
-                DetectAppsService.sendEvent(service, thisSample.getCategory(), "meaningfulness_scale", ";meaningfulness="+idx+";app="+thisSample.getPackageName()+";duration="+thisSample.getSampleTime());
+                //DetectAppsService.sendEvent(service, thisSample.getCategory(), thisSample.getPromptType(), ";meaningfulness="+idx+";app="+thisSample.getPackageName()+";duration="+thisSample.getSampleTime());
 
                 service.removeView(popup);
                 Random rand = new Random();
@@ -195,13 +281,20 @@ public class Sample {
                     final EditText answer = (EditText) followUp.findViewById(R.id.editText);
 
                     service.addView(followUp);
+                    thisSample.setPromptType("meaningfulness_text");
                     Button submit2 = (Button) followUp.findViewById(R.id.btnSubmit);
 
                     submit2.setOnClickListener(new View.OnClickListener() {
                         @Override
                         public void onClick(View v) {
-                            DetectAppsService.sendEvent(service, thisSample.getCategory(), "meaningfulness_text", ";response="+answer.getText()+";app="+thisSample.getPackageName()+";duration="+thisSample.getSampleTime());
-                            service.removeView(followUp);
+                            String response = answer.getText().toString();
+                            if (response.length() >= 75) {
+                                //DetectAppsService.sendEvent(service, thisSample.getCategory(), thisSample.getPromptType(), ";response="+answer.getText()+";app="+thisSample.getPackageName()+";duration="+thisSample.getSampleTime());
+                                service.removeView(followUp);
+                                final View followUp = thisSample.getPurpose(service);
+                                service.addView(followUp);
+                                thisSample.setPromptType("purpose_mc");
+                            }
                         }
                     });
                 }
